wpa2 si: add special handle for 4.2.3
The test plan only mentioned:
PBAC (MFPR and MFPC subfields shall be set to 1 if PBAC subfield is set
to 1).

However, the MFPC/MFPR are required to be 0 when PBAC in the test 4.2.3
by Allion lab. 88w8987 doesn't support PBAC but We can't always set
MFPC/MFPR because WPA3 requires MFPC/MFPR to be 1.

Apply the special handle for WPA2 SI 4.2.3 for passing both WPA2 SI and
WPA3 test plan.

Change-Id: I0375718ba8aa095ac58a6c6329460918dfde173c