A-MPDU: Modify aggrpriotbl for 11n 4.2.40
The old aggrpriotbl on LWIN21 was:
"2 2 0 0 1 1 3 3 4 4 5 5 255 255 255 255"

NXP suggest change to
"2 2 0 0 1 1 3 3 4 4 5 5 6 6 255 7" to pass 11n 4.2.40.

If not set in this way, the 11n 4.2.40 would fail with the following
message:
sniffer_frame_check,name,wireShark_wlan,filename,4.2.40_sniffer,\
Rxmac,00:50:43:02:FF:01,Txmac,00:10:18:F8:DE:08,framename,BA,Present,Yes

NXP sussgest only set it for 11n 4.2.40 so we add ch_11n_4_2_40_AMPDU.sh
to set the aggrpriotbl only when ssid is "Iq9865hfg|?" which is defined
for 11n 4.2.40 in the test plan.

Change-Id: Ia476156056731147139544b87e0e30017603ade5